Moved versus socket to lib folder -- working on a testing script but not sure...
