+++ lib/VersusSocket.js
@@ -10,7 +10,7 @@ const ROOT_DOMAIN = 'https://404jkfoundit.com'
 
 class VersusSocket extends DefaultClient {
   /** @override */
-  _getConnectURL(code: string): string {
+  _getConnectURL(code) {
     const endpoint = code ? `/versus/${code}/join` : '/versus/create'
     return ROOT_DOMAIN.replace('http', 'ws') + endpoint
   }
