+++ lib/versusEvents.js
