+++ package.json
@@ -2,6 +2,7 @@
   "name": "mathattack",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
     "android": "react-native run-android",
     "ios": "react-native run-ios",
@@ -43,7 +44,8 @@
     "react-native-svg-transformer": "^1.0.0",
     "react-test-renderer": "17.0.2",
     "webpack": "^5.65.0",
-    "webpack-cli": "^4.9.1"
+    "webpack-cli": "^4.9.1",
+    "ws": "^8.6.0"
   },
   "jest": {
     "preset": "react-native"
