+++ scenes/VersusFlow/ResultsAndPlayAgain.js
@@ -3,7 +3,7 @@ import {StyleSheet, View} from 'react-native'
 import {FullScreenOverlay, ScreenContainer} from '../../styles/elements'
 import TitleText from '../../components/TitleText'
 import {black, neonGreen, dimmedGreen, white} from '../../styles/colors'
-import VersusSocket from '../../models/VersusSocket'
+import VersusSocket from '../../lib/VersusSocket'
 import PropTypes from 'prop-types'
 import {EVENT_MarkReady, WON_FLAG_WON} from '../../constants/versus'
 import NormalText from '../../components/NormalText'
