+++ scenes/VersusFlow/VersusRound.js
@@ -5,7 +5,7 @@ import CalculatorInput from '../../components/UI/CalculatorInput'
 import {useDispatch, useSelector} from 'react-redux'
 import {selectUserAnswer} from '../../redux/selectors'
 import GameQuestion from '../../models/GameQuestion'
-import VersusSocket from '../../models/VersusSocket'
+import VersusSocket from '../../lib/VersusSocket'
 import PropTypes from 'prop-types'
 import {neonBlue, dimmedBlue} from '../../styles/colors'
 import EquationAndAnswerInterface from '../../components/UI/EquationAndAnswerInterface'
