+++ scenes/VersusFlow/WaitingForOpponent.js
@@ -6,7 +6,7 @@ import MenuButton from '../../components/MenuButton'
 import StringInput from '../../components/StringInput'
 import DividerLine from '../../components/DividerLine'
 import {spaceDefault} from '../../styles/layout'
-import VersusSocket from '../../models/VersusSocket'
+import VersusSocket from '../../lib/VersusSocket'
 import PropTypes from 'prop-types'
 import useCountdown from '../../hooks/useCountdown'
 import TitleText from '../../components/TitleText'
